# Faster inputTensor()