This is an initial attempt to fix #1143, following discussion with @neubig and @yoavg . By bypassing a few intermediate calls, I was able to reduce the time needed to input a 100000000-sized vector as follows:

CPU:
125.5sec -> 12.3sec
GPU:
13.5sec -> 6.5sec

The Expression returned by the new version no longer supports set(), so it cannot be used with static graphs. The question here would be how we handle this API-wise. As of this current implementation, we distinguish between the two cases via ```dy.inputExpr(np_arr)``` for dynamic graphs, and ```dy.inputExpr(np_arr,reusable_expr=True)``` for static graphs.

To-Do list:

- [ ] establish API
- [x] check if further speed-ups are possible
- [x] if needed, implement inputTensorTranspose()
- [x] practice test