@msperber 2018-01-24
This would be ready for a review now @neubig @yoavg . I tested it with a model that takes speech inputs and get better speed and reasonable outputs. I also tried the static-graph example from the documentation to make sure I didn't break anything there. I tried to keep API changes at a minimum, but this PR will break existing code that uses static graphs and inputTensor (in order to unbreak, people will have to add ```reusable_expr=True``` as argument). Alternatively, we could make True the default, but that will limit the speed-up experienced by users unless they add ```reusable_expr=False``` in their code.

Changes are summarized as follows:
- avoid some unnecessary intermediate calls
- for flat inputs, skip call to np.flatten()
- for non-reusable inputs, directly use float-vectors rather than the rather slow FloatVectorValue container
- best speed is achieved by the new inputTensorTranspose() operator that takes in a flat numpy array and desired dimensions and will directly pass it through in column-major format.
- deactivate bound checks

As before, we make sure things are not garbage-collected prematurely by referencing the stored values in the returned Expression which is itself held in the computation graph. I have a feeling that more speed-up might be possible, but this is about as far as I got without spending too much time digging into Cython memory management.